Fix moved file paths
